Move alternate builder form parameterization into BaselineTests
Keep all convention-driven, file-based test parameterization in one
place. The assertion stays at the parse level (both files must write
identically) rather than running the full pipeline: Vgo only writes
output that shrinks the file, and re-optimizing a baseline's rounded
coordinates can find conversions the original unrounded input did not,
so byte-for-byte pipeline comparison is not achievable for these
assets.
